requirements.txt
Pillow==9.2.0
Pillow==10.2.0
streamlit==1.31.1
